14-08-25 [PATCH 3] Cygwin compatibility
All makefiles are now compatible with cygwin and msys2.

For this, the files that are dependent on $(OS) and that are
included in the main makefile were created.

 ** scripts/windows-platform.mk -> window specific variables

This exports its variables via export to all makefiles marked
as $(SUBDIRS) via the $(MAKE) command.

 ** scripts/command.mk -> gcc, ld, ar, etc commands

To be compatible with Cygwin, MC now supports X11 and OpenGL3.3

* ZLIB now compiles with cmake (avoid use of native makefiles)
* Also fixed the level update reset (now resets to 0 every second)

In include/core/Host/Windows/exit.h the definition MC_INTERNAL_BREAK_POINT
was established depending on the Windows toolchain

 ** raise(SIGTRAP) for *emulated* unix enviroments (cygwin, msys2...)
 ** __debugbreak() for VS and mingw compilers

Now all comments will have the original author's signature at the
end, like this:

 ** Signed-off-by: author <email>

Signed-off-by: Andres26  <andrescamorao9@gmail.com>